Fix some more compiler warnings.
* libgda/gda-data-model.c: (gda_data_model_set_values): Fix a
  signed/unsigned warning.
* libgda-ui/data-entries/gdaui-entry-bin.c: (popup_position):
  Remove the call to
  gtk_widget_size_request() because the result is not used.
* libgda-ui/data-entries/widget-embedder.c:
  (widget_embedder_size_allocate): Remove the call to
  gtk_widget_get_child_requisition() because the result is not used.

+++ libgda/gda-data-model.c
@@ -780,7 +780,7 @@ gda_data_model_set_values (GdaDataModel *model, gint row, GList *values, GError 
 		/* save the values */
 		gint col, ncols;
 		ncols = gda_data_model_get_n_columns (model);
-		if (g_list_length (values) > ncols) {
+		if ((gint)g_list_length (values) > ncols) {
 			g_set_error (error, GDA_DATA_MODEL_ERROR,  GDA_DATA_MODEL_VALUES_LIST_ERROR,
 				     "%s", _("Too many values in list"));
 			return FALSE;
